[IMP] account_payment_order: No error on unknown payment method codes
Since v14, the tool for changing the target transfer account is to
create new payment methods, and put a method line in the journal with
such account.

Thus, there will be "fake" payment methods with no export handler
with the only goal of having a different target account.

Raising an error on such methods when finishing the payment order, forcing
to add custom code for returning (False, False), is too much, so better
to simply return that by default and don't block in these cases.

TT47723